requirements.txt
Django==3.0.8
djangorestframework==3.11.0
celery==4.4.6
requests==2.22.0
numpy==1.19.0
drf-yasg==1.17.0